chore(avc,mpeg2-codec,decode): clean up all clippy warnings, fix 6 tautological test assertions
bitvue-avc: 182 warnings/errors (123 warnings + 6 hard compile errors that
were blocking 3 test binaries from compiling at all -- fixing those errors
let the 3 binaries build for the first time, surfacing ~53 more warnings
that had been invisible until then) down to 0.

bitvue-mpeg2-codec: 7 down to 0 (2 collapsible_match in src/lib.rs, both
verified to preserve the trailing wildcard-arm fallthrough; 5 mechanical in
tests).

bitvue-decode: 15 down to 0 (5 in src/ -- redundant_closure in decoder.rs,
len_zero, 3x default_constructed_unit_structs in the SIMD strategy files --
and 10 across 3 test files).

overlay_extraction.rs (site of the real B-slice-fallback fix earlier this
session, commit abf832f) and slice.rs got only syntax-preserving changes
near sensitive code, each verified against the real match/branch structure
before editing -- no changes to mb_type classification or bit-reading logic.

6 real tautological assertions fixed (found while investigating "assert_eq!
with a literal bool" / "always true" warnings -- verified actual behavior
with debug prints before writing each replacement, matching the standard
already set by this session's earlier yuv_test.rs/leb128 fixes):
- overlay_extraction_test.rs (3x): `grid_w >= 0`/`coded_width >= 0` on
  unsigned fields -> real expected values from the fixture's SPS dimensions.
- frames_extraction_test.rs/frames_edge_cases_test.rs (3x): `len() >= 0` ->
  real observed counts (0 for incomplete-byte NALs, 1 for a zero-length
  frame between two start codes -- confirmed by running each test, not
  guessed).
- lib_api_test.rs: `!x.is_empty() || x.len() == 0` (a literal `A || !A`,
  always true) -> `assert_eq!(len, 0)` with the real reason (truncated SEI
  NAL has no length field to parse).

One deliberate exception: error_comprehensive_test.rs's
test_result_ok_variant triggers unnecessary_literal_unwrap on `Ok(42).unwrap()`,
but that's a genuine Result<T> alias smoke test, not a tautology -- kept with
a scoped, documented #[allow].

cargo build + test --lib --tests clean for all three crates (217 avc + 35
mpeg2 + 181 decode tests, 0 failures). clippy --lib --tests: 0 warnings in
all three (confirmed by filtering to each crate's own file paths). fmt
--check clean.

+++ crates/bitvue-avc/tests/integration_test.rs
@@ -275,7 +275,7 @@ fn test_frame_extraction_count() {
     // All extracted frames should have valid offsets
     for frame in &frames {
         assert!(
-            frame.offset as usize <= data.len(),
+            frame.offset <= data.len(),
             "Frame offset should be within data"
         );
         assert!(frame.size > 0, "Frame should have non-zero size");
